Move plan form to its own component
